DOMA-3099 extracted using `TaskLauncher` into `useTicketExportTask`
